requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-chess
pydantic
SQLAlchemy
psycopg2-binary
openai>=1.0.0
nanoid
numpy
passlib[bcrypt]
python-jose[cryptography]
